+++ src/js/records/entities/customer.record.js
@@ -18,6 +18,7 @@ export class CustomerRecord {
     this.firstName = _get(record, 'firstName')
     this.lastName = _get(record, 'lastName')
     this.status = _get(record, 'status')
+    this.addedAt = _get(record, 'addedAt')
     this.balances = _get(record, 'balances', [])
       .map(item => ({
         id: item.id,